+++ projects/ngrx-auto-entity/src/lib/actions/entity-info.ts
@@ -1,4 +1,4 @@
-import { IEntityNames, IEntityTransformer } from '../decorators/entity';
+import { IEntityNames, IEntityTransformer } from '../decorators/entity-options';
 
 /**
  * Descriptor of an Entity model and related metadata.
