+++ projects/ngrx-auto-entity/src/lib/decorators/entity-util.ts
@@ -1,24 +1,35 @@
+import { pipe } from '../../util/func';
 import { TNew } from '../actions/model-constructor';
-import { EntityComparer, IEntityOptions, IEntityTransformer } from './entity';
+import { EntityComparer, IEntityOptions, IEntityTransformer } from './entity-options';
 import { ENTITY_OPTS_PROP } from './entity-tokens';
 
-export const nameOfEntity = <TModel>(entityOrType: TNew<TModel> | TModel): string | null | undefined =>
-  entityOrType &&
-  ((entityOrType[ENTITY_OPTS_PROP] ||
-    (entityOrType.constructor ? entityOrType.constructor[ENTITY_OPTS_PROP] : {}) ||
-    {}) as IEntityOptions).modelName;
+export const EMPTY_OBJECT = {};
+
+export const getEntity = <TModel>(entityOrType: TNew<TModel> | TModel | TModel[]): TNew<TModel> | TModel =>
+  Array.isArray(entityOrType) ? entityOrType[0] : entityOrType;
+
+export const ensureObject = value => value || EMPTY_OBJECT;
+
+export const getEntityOptions = <TModel>(entityOrType: TNew<TModel> | TModel | TModel[]): IEntityOptions =>
+  (entityOrType[ENTITY_OPTS_PROP] ||
+    (entityOrType.constructor ? entityOrType.constructor[ENTITY_OPTS_PROP] : EMPTY_OBJECT) ||
+    EMPTY_OBJECT) as IEntityOptions;
+
+export const entityOptions = <TModel>(entityOrType: TNew<TModel> | TModel | TModel[]): IEntityOptions =>
+  pipe(
+    getEntity,
+    ensureObject,
+    getEntityOptions
+  )(entityOrType);
+
+export const nameOfEntity = <TModel>(entityOrType: TNew<TModel> | TModel): string | undefined =>
+  entityOptions(entityOrType).modelName;
 
 export const uriNameOfEntity = <TModel>(entityOrType: TNew<TModel> | TModel): string | null | undefined =>
-  entityOrType &&
-  ((entityOrType[ENTITY_OPTS_PROP] ||
-    (entityOrType.constructor ? entityOrType.constructor[ENTITY_OPTS_PROP] : {}) ||
-    {}) as IEntityOptions).uriName;
+  entityOptions(entityOrType).uriName;
 
 export const pluralNameOfEntity = <TModel>(entityOrType: TNew<TModel> | TModel): string | null | undefined =>
-  entityOrType &&
-  ((entityOrType[ENTITY_OPTS_PROP] ||
-    (entityOrType.constructor ? entityOrType.constructor[ENTITY_OPTS_PROP] : {}) ||
-    {}) as IEntityOptions).pluralName;
+  entityOptions(entityOrType).pluralName;
 
 export const mapComparer = (options: IEntityOptions, name: string): EntityComparer =>
   !!options.comparers
@@ -40,26 +51,14 @@ export const namedComparer = (options: IEntityOptions, name: string): EntityComp
 export const getComparer = (options: IEntityOptions, name?: string): EntityComparer =>
   !!options ? (!!name ? namedComparer(options, name) : defaultComparer(options)) : undefined;
 
-export const getEntity = <TModel>(entityOrType: TNew<TModel> | TModel | TModel[]): TNew<TModel> | TModel =>
-  Array.isArray(entityOrType) ? entityOrType[0] : entityOrType;
-
 export const entityComparer = <TModel>(
   entityOrType: TNew<TModel> | TModel | TModel[],
   name?: string
-): EntityComparer | null | undefined =>
-  (entityOrType = getEntity(entityOrType)) &&
-  entityOrType &&
-  getComparer(
-    (entityOrType[ENTITY_OPTS_PROP] ||
-      (entityOrType.constructor ? entityOrType.constructor[ENTITY_OPTS_PROP] : {}) ||
-      {}) as IEntityOptions,
-    name
-  );
+): EntityComparer | null | undefined => getComparer(entityOptions(entityOrType), name);
 
 export const entityTransforms = <TModel>(
   entityOrType: TNew<TModel> | TModel
-): IEntityTransformer[] | null | undefined =>
-  entityOrType &&
-  ((entityOrType[ENTITY_OPTS_PROP] ||
-    (entityOrType.constructor ? entityOrType.constructor[ENTITY_OPTS_PROP] : {}) ||
-    {}) as IEntityOptions).transform;
+): IEntityTransformer[] | null | undefined => entityOptions(entityOrType).transform;
+
+export const entityMaxAge = <TModel>(entityOrType: TNew<TModel> | TModel | TModel[]): number =>
+  entityOptions(entityOrType).defaultMaxAge;
